Just to avoid any database delete problems. Just to make sure that will NEVER happen :-)

+++ application/models/grocery_crud_model.php
@@ -354,6 +354,11 @@ class grocery_CRUD_Model  extends CI_Model  {
     function db_delete($primary_key_value)
     {
     	$primary_key_field = $this->get_primary_key();
+    	
+    	if($primary_key_field === false)
+    		return false;
+    	
+    	$this->db->limit(1);
     	$this->db->delete($this->table_name,array( $primary_key_field => $primary_key_value));
     	if( $this->db->affected_rows() != 1)
     		return false;
